Library/ETLAssist/requirements.txt
arcetl
pyodbc
python-dateutil
